create modules for load and query phases
